Fix noise-gate sign-wrap bug and stored XSS in config page
Found during an independent pre-publish code review.

- audio_dsp.c: the noise gate narrowed to int16_t before computing the
  absolute value, which overflows when a sample clamps to exactly -32768
  (a full-scale clipped peak) and wraps negative -- causing the gate to
  silently mute the loudest samples regardless of the configured threshold.
  Fixed by keeping the absolute-value computation in int32_t.

- web_server.c: WiFi SSID/password were reflected into the config page's
  HTML unescaped, unlike device_name (which is already character-restricted
  on input). Since POST /save has no auth/CSRF protection, this was a real
  stored-XSS vector. Added html_escape() and use it for all three free-text
  fields when rendering the page -- output encoding, not input filtering,
  since SSIDs/passwords can legitimately contain any printable character.

Both fixes verified in isolation via standalone test harnesses (not by
writing test payloads into the live device's actual WiFi config).

+++ src/audio_dsp.c
@@ -32,10 +32,14 @@ void audio_dsp_process(audio_dsp_state_t *st, int16_t *buf, size_t count)
         if (s >  32767) s =  32767;
         if (s < -32768) s = -32768;
 
-        // Noise gate: mute samples below threshold
+        // Noise gate: mute samples below threshold. Compute the absolute
+        // value in int32_t (s's own type) before comparing — narrowing to
+        // int16_t first overflows when s is exactly -32768 (a full-scale
+        // clipped sample), wrapping the result negative and causing the
+        // gate to fire on every clipped sample regardless of threshold.
         if (g_config.noise_gate > 0) {
-            int16_t a = (int16_t)(s < 0 ? -s : s);
-            if (a < (int16_t)g_config.noise_gate) s = 0;
+            int32_t a = s < 0 ? -s : s;
+            if (a < (int32_t)g_config.noise_gate) s = 0;
         }
 
         buf[i] = (int16_t)s;
